+++ src/azure-cli/azure/cli/command_modules/resource/aaz/latest/policy/assignment/non_compliance_message/_update.py
@@ -54,7 +54,7 @@ class Update(AAZCommand):
             help={"short-summary": "The name of the policy assignment.", "long-summary": "The name of the policy assignment is the name segment of its resource ID."},
             required=True,
             fmt=AAZStrArgFormat(
-                pattern="^[^<>*%&:\\?.+/]*[^<>*%&:\\?.+/ ]+$",
+                pattern="^[^<>*%&:\\?+/]*[^<>*%&:\\?+/ ]+$",
             ),
         )
         _args_schema.scope = AAZStrArg(