+++ src/azure-cli/azure/cli/command_modules/resource/aaz/latest/policy/definition/_show.py
@@ -59,7 +59,7 @@ class Show(AAZCommand):
             help="The name of the policy definition.",
             required=True,
             fmt=AAZStrArgFormat(
-                pattern="^[^<>*%&:\\?.+/]*[^<>*%&:\\?.+/ ]+$",
+                pattern="^[^<>*%&:\\?+/]*[^<>*%&:\\?+/ ]+$",
             ),
         )
         return cls._args_schema