+++ src/azure-cli/azure/cli/command_modules/resource/aaz/latest/policy/set_definition/_create.py
@@ -65,7 +65,7 @@ class Create(AAZCommand):
             help="The name of the policy set definition.",
             required=True,
             fmt=AAZStrArgFormat(
-                pattern="^[^<>*%&:\\?.+/]*[^<>*%&:\\?.+/ ]+$",
+                pattern="^[^<>*%&:\\?+/]*[^<>*%&:\\?+/ ]+$",
             ),
         )