+++ src/azure-cli/azure/cli/command_modules/resource/aaz/latest/policy/set_definition/_delete.py
@@ -56,7 +56,7 @@ class Delete(AAZCommand):
             help="The name of the policy set definition.",
             required=True,
             fmt=AAZStrArgFormat(
-                pattern="^[^<>*%&:\\?.+/]*[^<>*%&:\\?.+/ ]+$",
+                pattern="^[^<>*%&:\\?+/]*[^<>*%&:\\?+/ ]+$",
             ),
         )
         return cls._args_schema